skeletons/aqua: Separate into modules
